cleaned up methodes (multi file approach)
