+++ api/database_errors.go
@@ -1,0 +1,12 @@
+package api
+
+import (
+	"net/http"
+)
+
+var (
+	ErrPersonNotFound = APIError{
+		Message: "Person Not Found",
+		Status:  http.StatusNotFound,
+	}
+)
