+++ api/phone_methodes.go
@@ -1,0 +1,95 @@
+package api
+
+import (
+	"fmt"
+	"github.com/nyaruka/phonenumbers"
+	"github.com/sundowndev/phoneinfoga/v2/lib/number"
+	"log"
+	"net/http"
+	"strings"
+)
+
+// Markdown
+func (phoneNumber PhoneNumber) Markdown() string {
+	var sb strings.Builder
+	if phoneNumber.IsValid() {
+		sb.WriteString(fmt.Sprintf("- Phone: `%s`\n", phoneNumber.Number))
+	}
+	return sb.String()
+}
+
+func (numbers PhoneNumbers) Markdown() string {
+	var sb strings.Builder
+	for _, number := range SortMapKeys(map[string]PhoneNumber(numbers)) {
+		sb.WriteString(numbers[number].Markdown())
+	}
+	return sb.String()
+}
+
+// Validation
+
+func (phoneNumber PhoneNumber) IsValid() bool {
+	parsedNumber, err := phonenumbers.Parse(phoneNumber.Number, "")
+	if err != nil {
+		log.Printf("error parsing number: %s", err)
+		return false
+	}
+	return phonenumbers.IsValidNumber(parsedNumber)
+}
+
+func (numbers PhoneNumbers) Validate() error {
+	for _, number := range SortMapKeys(map[string]PhoneNumber(numbers)) {
+		if number != numbers[number].Number {
+			return APIError{
+				Message: fmt.Sprintf("Key missmatch: Phone[%s] = %s", number, numbers[number].Number),
+				Status:  http.StatusBadRequest,
+			}
+		}
+	}
+	return nil
+}
+
+// Parsing
+
+func (phoneNumber PhoneNumber) Parse() (PhoneNumber, error) {
+	phoneNumber.Valid = phoneNumber.IsValid()
+	if phoneNumber.Valid {
+		parsedNumber, err := phonenumbers.Parse(phoneNumber.Number, "")
+		if err != nil {
+			log.Printf("error parsing number: %s", err)
+			return phoneNumber, err
+		}
+		phoneNumber.Number = phonenumbers.Format(parsedNumber, phonenumbers.INTERNATIONAL)
+		phoneNumber.NationalFormat = phonenumbers.Format(parsedNumber, phonenumbers.NATIONAL)
+		phoneNumber.Phoneinfoga, err = phoneNumber.GetPhoneinfoga() // FIXME error handeling
+		if err != nil {
+			log.Printf("error getting number number: %s", err)
+			return phoneNumber, err
+		}
+	}
+	return phoneNumber, nil
+}
+
+func (numbers PhoneNumbers) Parse() (PhoneNumbers, error) {
+	newNumbers := PhoneNumbers{}
+	for _, number := range SortMapKeys(map[string]PhoneNumber(numbers)) {
+		// delete empty phone numbers
+		if number == "" {
+			delete(numbers, number) // number == ""
+			break
+		}
+		parsedNumber, err := numbers[number].Parse()
+		if err != nil {
+			return newNumbers, err
+		}
+		newNumbers[parsedNumber.Number] = parsedNumber
+	}
+	return newNumbers, nil
+}
+
+// Info
+
+func (phoneNumber PhoneNumber) GetPhoneinfoga() (number.Number, error) {
+	n, err := number.NewNumber(phoneNumber.Number)
+	return *n, err
+}
